+++ controllers/ConfigController.php
@@ -4,7 +4,6 @@ namespace humhub\modules\flexTheme\controllers;
 
 use humhub\modules\flexTheme\models\Config;
 use humhub\modules\flexTheme\models\ColorSettings;
-use humhub\modules\flexTheme\models\DarkMode;
 use humhub\modules\flexTheme\models\AdvancedSettings;
 use Yii;
 
@@ -34,7 +33,10 @@ class ConfigController extends \humhub\modules\admin\components\Controller
         $form = new ColorSettings();
 
         if ($form->load(Yii::$app->request->post()) && $form->save()) {
+            Yii::$app->assetManager->clear();
+            Yii::$app->view->theme->variables->flushCache();
             $this->view->saved();
+            return $this->redirect(['/flex-theme/config/colors']);
         }
 
         return $this->render('colors', ['model' => $form]);
@@ -42,13 +44,7 @@ class ConfigController extends \humhub\modules\admin\components\Controller
 
     public function actionDarkMode()
     {
-        $form = new DarkMode();
-
-        if ($form->load(Yii::$app->request->post()) && $form->save()) {
-            $this->view->saved();
-        }
-
-        return $this->render('dark-mode', ['model' => $form]);
+        return $this->render('dark-mode');
     }
 
     public function actionAdvanced()
@@ -56,14 +52,12 @@ class ConfigController extends \humhub\modules\admin\components\Controller
         $form = new AdvancedSettings();
         $config = new Config();
         $colorSettings = new ColorSettings();
-        $darkModeSettings = new DarkMode();
 
         if(!empty(Yii::$app->request->post())) {
             $data = json_decode(Yii::$app->request->post()['AdvancedSettings']['settingsJson'], true);
 
             $config->load($data);
             $colorSettings->load($data);
-            $darkModeSettings->load($data);
 
             // Check validation
             if (!$config->validate()) {
@@ -74,13 +68,12 @@ class ConfigController extends \humhub\modules\admin\components\Controller
                 $form->addError('settingsJson', Yii::t('FlexThemeModule.admin', 'There seem to be invalid values!') . ' (ColorSettings)');
 			    return $this->render('advanced', ['model' => $form]);
             }
-            if (!$darkModeSettings->validate()) {
-                $form->addError('settingsJson', Yii::t('FlexThemeModule.admin', 'There seem to be invalid values!') . ' (DarkModeSettings)');
-			    return $this->render('advanced', ['model' => $form]);
-            }
             // Save
-            if ($config->save() && $colorSettings->save() && $darkModeSettings->save() && $form->load(Yii::$app->request->post()) && $form->save()) {
+            if ($config->save() && $colorSettings->save() && $form->load(Yii::$app->request->post()) && $form->save()) {
+                Yii::$app->assetManager->clear();
+                Yii::$app->view->theme->variables->flushCache();
                 $this->view->saved();
+                return $this->redirect(['/flex-theme/config/advanced']);
             }
         }