+++ models/AdvancedSettings.php
@@ -4,7 +4,6 @@ namespace humhub\modules\flexTheme\models;
 
 use humhub\modules\flexTheme\models\Config;
 use humhub\modules\flexTheme\models\ColorSettings;
-use humhub\modules\flexTheme\models\DarkMode;
 
 use Yii;
 
@@ -63,7 +62,6 @@ class AdvancedSettings extends \yii\base\Model
     {
         $config = [];
         $colors = [];
-        $darkMode = [];
 
         $module = Yii::$app->getModule('flex-theme');
 
@@ -89,22 +87,6 @@ class AdvancedSettings extends \yii\base\Model
             }
         }
 
-        // Get Dark Mode settings
-        $darkModeEnabled = $module->settings->get('darkModeEnabled');
-        if (!empty($darkModeEnabled)) {
-            $darkMode['darkModeEnabled'] = (bool) $darkModeEnabled;
-        }
-        // Get dark colors
-        $color_names = array_merge(DarkMode::MAIN_COLORS, DarkMode::TEXT_COLORS, DarkMode::BACKGROUND_COLORS);
-        foreach( $color_names as $color ) {
-            $value = $module->settings->get('dark_' . $color);
-
-            // exclude empty settings
-            if(!empty($value)) {
-                $darkMode[$color] = $value;
-            }
-        }
-
-        return ['Config' => $config, 'ColorSettings' => $colors, 'DarkMode' => $darkMode];
+        return ['Config' => $config, 'ColorSettings' => $colors];
     }
 }