+++ views/config/dark-mode.php
@@ -1,41 +1,15 @@
 <?php
+use yii\helpers\Url;
 
-use humhub\modules\flexTheme\models\DarkMode;
-use humhub\modules\ui\form\widgets\ActiveForm;
-use kartik\widgets\ColorInput;
-use yii\helpers\Html;
-
-$main_colors = DarkMode::MAIN_COLORS;
-$text_colors = DarkMode::TEXT_COLORS;
-$background_colors = DarkMode::BACKGROUND_COLORS;
+$darkModeUrl = Url::toRoute('/marketplace/browse?keyword=dark-mode&tags=');
+$darkModeName = Yii::t('FlexThemeModule.admin', 'Dark Mode');
+$link = '<a href=' . $darkModeUrl . '>' . $darkModeName . '</a>';
 ?>
 <div class="panel-body">
-    <?php $form = ActiveForm::begin(['id' => 'configure-form']);?>
-
-        <?= $form->field($model, 'darkModeEnabled')->checkbox(); ?>
-
-        <?= $form->beginCollapsibleFields(Yii::t('FlexThemeModule.admin', 'Main Colors')); ?>
-            <?php foreach ($main_colors as $color): ?>
-                <?= $form->field($model, $color)->widget(ColorInput::class, ['options' => ['placeholder' => \Yii::t('FlexThemeModule.admin', 'Select color ...')]]); ?>
-            <?php endforeach; ?>
-        <?= $form->endCollapsibleFields(); ?>
-
-        <?= $form->beginCollapsibleFields(Yii::t('FlexThemeModule.admin', 'Text Colors')); ?>
-            <?php foreach ($text_colors as $color): ?>
-                <?= $form->field($model, $color)->widget(ColorInput::class, ['options' => ['placeholder' => \Yii::t('FlexThemeModule.admin', 'Select color ...')]]); ?>
-            <?php endforeach; ?>
-        <?= $form->endCollapsibleFields(); ?>
-
-        <?= $form->beginCollapsibleFields(Yii::t('FlexThemeModule.admin', 'Background Colors')); ?>
-            <?php foreach ($background_colors as $color): ?>
-                <?= $form->field($model, $color)->widget(ColorInput::class, ['options' => ['placeholder' => \Yii::t('FlexThemeModule.admin', 'Select color ...')]]); ?>
-            <?php endforeach; ?>
-        <?= $form->endCollapsibleFields(); ?>
-
-    <div class="form-group">
-        <?= Html::submitButton(Yii::t('base', 'Save'), ['class' => 'btn btn-primary', 'data-ui-loader' => '']) ?>
+    <div class="alert alert-info">
+        <p><?= Yii::t('FlexThemeModule.admin', 'Please use the module {darkmode} and select "HumHub (dark)".', [
+            'darkmode' => $link,
+        ]) ?>
+        </p>
     </div>
-
-    <?php ActiveForm::end(); ?>
-
 </div>